Move partial_model tests
